Fix various warnings across engine

+++ plugins/Grindstone.Renderables.3D/source/Mesh3dRenderer.cpp
@@ -177,7 +177,7 @@ static bool IsInFrustum(const CullingFrustum& frustum, const glm::mat4& viewMode
 }
 
 static CullingFrustum CreateFrustum(const Grindstone::Rendering::RenderViewData& renderViewData) {
-	float aspectRatio = renderViewData.renderArea.extent.x / renderViewData.renderArea.extent.y;
+	float aspectRatio = static_cast<float>(renderViewData.renderArea.extent.x) / static_cast<float>(renderViewData.renderArea.extent.y);
 	float tanFov = 1.0f / renderViewData.projectionMatrix[0][0];
 
 	float projection_43 = renderViewData.projectionMatrix[3][2];
